+++ edx_proctoring/urls.py
@@ -131,13 +131,6 @@ urlpatterns = [
         views.ProctoredExamAttemptView.as_view(),
         name='proctored_exam.exam_attempts'
     ),
-    # TODO: remove url after updating frontend-lib-special-exams
-    url(
-        r'edx_proctoring/v1/proctored_exam/attempt/course_id/{}/content_id/{}$'.format(
-            settings.COURSE_ID_PATTERN, CONTENT_ID_PATTERN),
-        views.ProctoredExamAttemptView.as_view(),
-        name='proctored_exam.exam_attempts_old'
-    ),
     url(
         r'edx_proctoring/v1/proctored_exam/settings/exam_id/(?P<exam_id>\d+)/$',
         views.ProctoredSettingsView.as_view(),